rehouse param helper functions in triggers.py
