+++ bin/hdfcoinc/pycbc_fit_sngls_binned
@@ -26,6 +26,7 @@ import copy, numpy as np
 from scipy.stats import kstest
 
 from pycbc import io, events, bin_utils, results
+from pycbc.events import triggers
 from pycbc.events import trigger_fits as trstats
 from pycbc.events.stat import sngl_statistic_dict
 import pycbc.version
@@ -201,7 +202,7 @@ def get_pars(args, tag, m1, m2, s1z, s2z):
         logging.info('Getting %s values for %i triggers' % (paramarg, len(m1)))
     except:
         pass
-    return trstats.get_param(paramarg, args, m1, m2, s1z, s2z)
+    return triggers.get_param(paramarg, args, m1, m2, s1z, s2z)
 
 if args.prune_param:
     logging.info('Getting min and max param values')
@@ -241,7 +242,7 @@ if args.prune_param:
         lstat = statpruneall[loudest]
         ltid = tidpruneall[loudest]
         ltime = timepruneall[loudest]
-        m1, m2, s1z, s2z = trstats.get_masses(templatef, ltid)
+        m1, m2, s1z, s2z = triggers.get_mass_spin(templatef, ltid)
         lbin = trstats.which_bin(get_pars(args, 'prune', m1, m2, s1z, s2z),
                                  minprpar, maxprpar,
                                  args.prune_bins, log=args.log_prune_param)
@@ -284,7 +285,7 @@ if args.prune_param:
 if trig_dur:
     binpars = tdur + args.min_duration
 else:
-    m1, m2, s1z, s2z = trstats.get_masses(templatef, tid)
+    m1, m2, s1z, s2z = triggers.get_mass_spin(templatef, tid)
     binpars = get_pars(args, 'bin', m1, m2, s1z, s2z)
 logging.info("Parameter range of triggers: %f - %f" %
                                                   (min(binpars), max(binpars)))
