+++ bin/hdfcoinc/pycbc_fit_sngls_over_multiparam
@@ -15,7 +15,7 @@
 from __future__ import division
 
 import sys, h5py, argparse, logging, pycbc.version, numpy
-from pycbc.events import trigger_fits as trstats
+from pycbc.events import triggers
 
 parser = argparse.ArgumentParser(usage="",
     description="Smooth (regress) the dependence of coefficients describing "
@@ -80,12 +80,12 @@ tid = fits['template_id'][:]
 
 logging.info('Calculating template parameter values')
 bank = h5py.File(args.bank_file, 'r')
-m1, m2, s1z, s2z = trstats.get_masses(bank, tid)
+m1, m2, s1z, s2z = triggers.get_mass_spin(bank, tid)
 
 parvals = []
 
 for param, slog in zip(args.fit_param, args.log_param):
-    data = trstats.get_param(param, args, m1, m2, s1z, s2z)
+    data = triggers.get_param(param, args, m1, m2, s1z, s2z)
     if slog in ['false', 'False', 'FALSE']:
         logging.info('Using param: %s', param)
         parvals.append(data)
