+++ bin/plotting/pycbc_plot_trigrate
@@ -25,7 +25,7 @@ import copy, numpy as np
 from scipy import stats as scistats
 
 from pycbc import io, events, bin_utils, results
-from pycbc.events import trigger_fits as trstats
+from pycbc.events import triggers
 from pycbc.events.stat import sngl_statistic_dict
 import pycbc.version
 
@@ -184,7 +184,7 @@ if not args.gps_start_time:
 plottime = time - args.gps_start_time
 
 if args.thorne_limit:
-    m1, m2, s1z, s2z = trstats.get_masses(templatef, tid)
+    m1, m2, s1z, s2z = triggers.get_mass_spin(templatef, tid)
     inside = np.logical_and(abs(s1z) < 0.998, abs(s2z) < 0.998)
     #stat = stat[inside]
     tid = tid[inside]
@@ -200,13 +200,13 @@ def get_pars(args, tag, m1, m2, s1z, s2z):
         logging.info('Getting %s values for %i triggers' % (paramarg, len(m1)))
     except:
         pass
-    return trstats.get_param(paramarg, args, m1, m2, s1z, s2z)
+    return triggers.get_param(paramarg, args, m1, m2, s1z, s2z)
 
 # get binning params
 if trig_dur:
     binpars = tdur + args.min_duration
 else:
-    m1, m2, s1z, s2z = trstats.get_masses(templatef, tid)
+    m1, m2, s1z, s2z = triggers.get_mass_spin(templatef, tid)
     binpars = get_pars(args, 'bin', m1, m2, s1z, s2z)
 logging.info("Parameter range of triggers: %f - %f" %
                                                   (min(binpars), max(binpars)))
