+++ pycbc/events/trigger_fits.py
@@ -214,73 +214,6 @@ def KS_test(distr, vals, alpha, thresh=None):
     return kstest(vals, cdf_fn)
 
 
-def get_masses(bank, tid):
-    """
-    Helper function
-
-    Parameters
-    ----------
-    bank : h5py File object
-        Bank parameter file
-    tid : integer or array of int
-        Indices of the entries to be returned
-
-    Returns
-    -------
-    m1, m2, s1z, s2z : tuple of floats or arrays of floats
-        Parameter values of the bank entries
-    """
-    m1 = bank['mass1'][:][tid]
-    m2 = bank['mass2'][:][tid]
-    s1z = bank['spin1z'][:][tid]
-    s2z = bank['spin2z'][:][tid]
-    return m1, m2, s1z, s2z
-
-
-def get_param(par, args, m1, m2, s1z, s2z):
-    """
-    Helper function
-
-    Parameters
-    ----------
-    par : string
-        Name of parameter to calculate
-    args : Namespace object returned from ArgumentParser instance
-        Calling code command line options, used for f_lower value
-    m1 : float or array of floats
-        First binary component mass (etc.)
-
-    Returns
-    -------
-    parvals : float or array of floats
-        Calculated parameter values
-    """
-    if par == 'mchirp':
-        parvals, _ = pnutils.mass1_mass2_to_mchirp_eta(m1, m2)
-    elif par == 'mtotal':
-        parvals = m1 + m2
-    elif par =='eta':
-        parvals = conversions.eta_from_mass1_mass2(m1, m2)
-    elif par in ['chi_eff', 'effective_spin']:
-        parvals = conversions.chi_eff(m1, m2, s1z, s2z)
-    elif par == 'template_duration':
-        # default to SEOBNRv4 duration function
-        parvals = pnutils.get_imr_duration(m1, m2, s1z, s2z, args.f_lower,
-                                           args.approximant or "SEOBNRv4")
-        if args.min_duration:
-            parvals += args.min_duration
-    elif par == 'tau0':
-        parvals = conversions.tau0_from_mass1_mass2(m1, m2, args.f_lower)
-    elif par == 'tau3':
-        parvals = conversions.tau3_from_mass1_mass2(m1, m2, args.f_lower)
-    elif par in pnutils.named_frequency_cutoffs.keys():
-        parvals = pnutils.frequency_cutoff_from_name(par, m1, m2, s1z, s2z)
-    else:
-        # try asking for a LALSimulation frequency function
-        parvals = pnutils.get_freq(par, m1, m2, s1z, s2z)
-    return parvals
-
-
 def which_bin(par, minpar, maxpar, nbins, log=False):
     """
     Helper function
